Separation flipping card and its controls
