+++ src/components/card/Card.tsx
@@ -1,8 +1,8 @@
 import React, { HTMLAttributes, useState } from 'react';
 import { MiniCard } from './MiniCard';
-import { FlippingCard } from './FlippingCard';
 import { CardType } from '@/state/slices';
 import { useAppStore } from '@/state';
+import { ZoomedCard } from './ZoomedCard';
 
 interface CardProps extends HTMLAttributes<React.FC> {
   mode: 'train' | 'edit';
@@ -22,7 +22,7 @@ export const Card: React.FC<CardProps> = ({ mode, cardData, className }) => {
           setZoomed(true);
         }}
       />
-      <FlippingCard
+      <ZoomedCard
         isShown={zoomed}
         close={() => setZoomed(false)}
         mode={mode}
