+++ tailwind.config.js
@@ -1,41 +1,41 @@
 /** @type {import('tailwindcss').Config} */
-/** @type {import('tailwindcss/colors')} */
-
-const colors = require('tailwindcss/colors');
 
 // eslint-disable-next-line no-undef
 module.exports = {
   darkMode: 'class', // переводим в режим определения dark мода (срабатывают dark: модификаторы) по наличию класса dark у родителя
   content: ['./index.html', './src/**/*.{js,ts,jsx,tsx}'],
   theme: {
-    colors: {
-      ...colors,
-      '1-1': '#178058',
-      '1-2': '#2d8b67',
-      '1-3': '#27a876',
-      '1-4': '#53BD94',
-      '1-5': '#77BFA3',
-      '1-6': '#98C9A3',
-      '1-7': '#BFD8BD',
-      '1-8': '#DDE7C7',
-      '1-9': '#EDEEC9',
-      '1-10': '#F5F0C4',
-      '1-11': '#FEF4D2',
-      '1-12': '#f8f3e4',
-      '2-1': '#02132A',
-      '2-2': '#0F233F',
-      '2-3': '#1D3456',
-      '2-4': '#2A5D6C',
-      '2-5': '#388A83',
-      '2-6': '#4CCAA4',
-      'o-black': '#02132A',
-      'o-red': '#E63946',
-      'o-white': '#F1FAEE',
-      'o-blue': '#A8DADC',
-      'o-mid-blue': '#457B9D',
-      'o-dark-blue': '#1D3557',
-    },
     extend: {
+      height: {
+        '1/12': '8.333333%',
+        '11/12': '91.666666%',
+      },
+      colors: {
+        '1-1': '#178058',
+        '1-2': '#2d8b67',
+        '1-3': '#27a876',
+        '1-4': '#53BD94',
+        '1-5': '#77BFA3',
+        '1-6': '#98C9A3',
+        '1-7': '#BFD8BD',
+        '1-8': '#DDE7C7',
+        '1-9': '#EDEEC9',
+        '1-10': '#F5F0C4',
+        '1-11': '#FEF4D2',
+        '1-12': '#f8f3e4',
+        '2-1': '#02132A',
+        '2-2': '#0F233F',
+        '2-3': '#1D3456',
+        '2-4': '#2A5D6C',
+        '2-5': '#388A83',
+        '2-6': '#4CCAA4',
+        'o-black': '#02132A',
+        'o-red': '#E63946',
+        'o-white': '#F1FAEE',
+        'o-blue': '#A8DADC',
+        'o-mid-blue': '#457B9D',
+        'o-dark-blue': '#1D3557',
+      },
       keyframes: {
         appear: {
           '0%': { transform: 'scale(0.62)', opacity: 0 },
